[refactor] Extension 분리
