+++ WeatherForecast/WeatherForecast/Extensions/CALayer+Extension.swift
@@ -7,24 +7,6 @@
 
 import UIKit
 
-extension UIView {
-    public static var spacerView: UIView {
-        let view = UIView()
-        view.isUserInteractionEnabled = false
-        view.setContentHuggingPriority(.required, for: .horizontal)
-        view.setContentCompressionResistancePriority(.dragThatCannotResizeScene, for: .horizontal)
-        return view
-    }
-}
-
-extension UIStackView {
-    func addArrangedSubviews(_ views: [UIView]) {
-        views.forEach {
-            self.addArrangedSubview($0)
-        }
-    }
-}
-
 extension CALayer {
     func addBorder(_ arr_edge: [UIRectEdge], color: UIColor, width: CGFloat) {
         for edge in arr_edge {
