+++ WeatherForecast/WeatherForecast/Extensions/Double+Extension.swift
